+++ examples/with-one-login-per-subdomain/api-server.js
@@ -5,6 +5,7 @@ let Session = require("supertokens-node/recipe/session");
 let { verifySession } = require("supertokens-node/recipe/session/framework/express");
 let { middleware, errorHandler } = require("supertokens-node/framework/express");
 let EmailPassword = require("supertokens-node/recipe/emailpassword");
+let Dashboard = require("supertokens-node/recipe/dashboard");
 
 const apiPort = process.env.REACT_APP_API_PORT || 3001;
 const apiDomain = process.env.REACT_APP_API_URL || `http://example.com:${apiPort}`;
@@ -42,6 +43,7 @@ supertokens.init({
             },
         }),
         Session.init(),
+        Dashboard.init(),
     ],
 });