+++ examples/with-thirdpartyemailpassword/api-server.js
@@ -55,9 +55,7 @@ supertokens.init({
             ],
         }),
         Session.init(),
-        Dashboard.init({
-            apiKey: "supertokens_is_awesome",
-        }),
+        Dashboard.init(),
     ],
 });