+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.31.3] - 2023-03-23
+
+### Changes
+
+-   Fix unsupported engine warning for npm v9
+-   Updates dependency versions for `supertokens-auth-react` and `supertokens-node` for all example apps
+-   Initialises the Dashboard recipe for all example apps
+
+## [0.31.2] - 2023-03-07
+
+### Fixes
+
+-   Keep redirectToPath's query params while redirecting to it from auth
+
+### Changes
+
+-   Update example apps after header based auth
+
 ## [0.31.1] - 2023-02-12
 
 ### Changes